updated copy commands
git-svn-id: svn://136.177.114.72/svn_GW/IPhreeqc/trunk@4361 1feff8c3-07ed-0310-ac33-dd36852eb9cd

+++ unit/TestIPhreeqc.cpp
@@ -1017,7 +1017,7 @@ void TestIPhreeqc::TestAddError(void)
 	// make sure initialized to empty
 	//
 	const char *expected = "TESTING AddError\n";
-	CPPUNIT_ASSERT_EQUAL(1u, obj.AddError(expected));
+	CPPUNIT_ASSERT_EQUAL((size_t)1, obj.AddError(expected));
 
 	// check 1
 	//
@@ -1027,7 +1027,7 @@ void TestIPhreeqc::TestAddError(void)
 	// check increment
 	//
 	const char *expected2 = "XXXXXX\n";
-	CPPUNIT_ASSERT_EQUAL(2u, obj.AddError(expected2));
+	CPPUNIT_ASSERT_EQUAL((size_t)2, obj.AddError(expected2));
 
 	// check concatenation
 	//
